backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Authentication app backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
]
